+++ NEWS.md
@@ -1,5 +1,7 @@
 # purrr (development version)
 
+* Added a new keep_empty argument to list_c(), list_cbind() and list_rbind() which will keep empty elements as NA in the returned dataframe. (#1096)
+
 # purrr 1.0.2
 
 * Fixed valgrind issue.